Move IReferrrable to Objects.GameState
