+++ FrEee.Core.Domain/Modding/Abilities/Ability.cs
@@ -7,7 +7,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using FrEee.Objects.GameState;
-using FrEee.Serialization;
 using FrEee.Utility;
 
 namespace FrEee.Modding.Abilities;
