+++ FrEee.Core.Domain/Modding/EventType.cs
@@ -1,5 +1,4 @@
 ﻿using FrEee.Objects.GameState;
-using FrEee.Serialization;
 using System.Collections.Generic;
 
 namespace FrEee.Modding;
