+++ FrEee.Core.Domain/Modding/Loaders/EventTypeLoader.cs
@@ -1,5 +1,4 @@
 ﻿using FrEee.Objects.GameState;
-using FrEee.Serialization;
 using System.Collections.Generic;
 using System.Linq;
