+++ FrEee.Core.Domain/Objects/Civilization/Construction/IConstructionTemplate.cs
@@ -1,7 +1,6 @@
 ﻿using FrEee.Objects.Civilization;
 using FrEee.Objects.GameState;
 using FrEee.Objects.Technology;
-using FrEee.Serialization;
 using FrEee.Utility;
 namespace FrEee.Objects.Civilization.Construction;
