+++ FrEee.Core.Domain/Objects/Civilization/Diplomacy/Messages/IMessage.cs
@@ -1,5 +1,4 @@
 ﻿using FrEee.Objects.GameState;
-using FrEee.Serialization;
 
 namespace FrEee.Objects.Civilization.Diplomacy.Messages;
