+++ FrEee.Core.Domain/Objects/Civilization/Orders/IOrder.cs
@@ -1,6 +1,5 @@
 ﻿using FrEee.Objects.GameState;
 using FrEee.Objects.LogMessages;
-using FrEee.Serialization;
 using System.Collections.Generic;
 
 namespace FrEee.Objects.Civilization.Orders;
