+++ FrEee.Core.Domain/Objects/Civilization/Orders/IRecyclable.cs
@@ -1,6 +1,5 @@
 ﻿using FrEee.Objects.GameState;
 using FrEee.Objects.Space;
-using FrEee.Serialization;
 using FrEee.Utility;
 namespace FrEee.Objects.Civilization.Orders;
