+++ FrEee.Core.Domain/Objects/Civilization/Race.cs
@@ -10,7 +10,6 @@ using System.IO;
 using System.Linq;
 using FrEee.Objects.GameState;
 using FrEee.Utility;
-using FrEee.Serialization;
 using FrEee.Modding.Abilities;
 
 namespace FrEee.Objects.Civilization;
