+++ FrEee.Core.Domain/Objects/Commands/CreateDesignCommand.cs
@@ -4,7 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using FrEee.Objects.Vehicles;
-using FrEee.Serialization;
+using FrEee.Objects.GameState;
 
 namespace FrEee.Objects.Commands;
