+++ FrEee.Core.Domain/Objects/Commands/CreateFleetCommand.cs
@@ -1,7 +1,6 @@
 ﻿using FrEee.Objects.Civilization;
 using FrEee.Objects.GameState;
 using FrEee.Objects.Space;
-using FrEee.Serialization;
 using System.Collections.Generic;
 
 namespace FrEee.Objects.Commands;
