+++ FrEee.Core.Domain/Objects/Commands/ICommand.cs
@@ -1,6 +1,5 @@
 using FrEee.Objects.Civilization;
 using FrEee.Objects.GameState;
-using FrEee.Serialization;
 using System.Collections.Generic;
 
 namespace FrEee.Objects.Commands;
