+++ FrEee.Core.Domain/Objects/GameState/GalaxyProgress.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using FrEee.Serialization;
 using FrEee.Utility;
 
 namespace FrEee.Objects.GameState
