+++ FrEee.Core.Domain/Objects/Space/Planet.cs
@@ -16,7 +16,6 @@ using FrEee.Objects.Civilization.Construction;
 using FrEee.Objects.Civilization.Orders;
 using FrEee.Objects.GameState;
 using FrEee.Extensions;
-using FrEee.Serialization;
 using FrEee.Utility;
 using FrEee.Processes.Combat;
 using FrEee.Modding.Abilities;
