+++ FrEee.Core.Domain/Objects/Space/StarSystem.cs
@@ -11,7 +11,6 @@ using FrEee.Serialization;
 using FrEee.Extensions;
 using FrEee.Objects.GameState;
 using FrEee.Utility;
-using FrEee.Serialization;
 using FrEee.Modding.Abilities;
 
 namespace FrEee.Objects.Space;
