+++ FrEee.Core.Domain/Objects/Space/WarpPoint.cs
@@ -5,7 +5,6 @@ using FrEee.Extensions;
 using System;
 using System.Linq;
 using FrEee.Objects.GameState;
-using FrEee.Serialization;
 using FrEee.Utility;
 using FrEee.Modding.Abilities;
