+++ FrEee.Core.Domain/Objects/Vehicles/IDesign.cs
@@ -5,7 +5,6 @@ using FrEee.Utility;
 using System.Collections.Generic;
 using FrEee.Objects.Civilization.Construction;
 using FrEee.Objects.Commands;
-using FrEee.Serialization;
 using FrEee.Objects.GameState;
 using FrEee.Modding.Abilities;
