+++ FrEee.Core.Domain/Objects/Vehicles/IVehicle.cs
@@ -6,9 +6,9 @@ using FrEee.Serialization;
 using FrEee.Utility;
 using System.Collections.Generic;
 using FrEee.Utility;
-using FrEee.Serialization;
 using FrEee.Processes.Combat;
 using FrEee.Modding.Abilities;
+using FrEee.Objects.GameState;
 
 namespace FrEee.Objects.Vehicles;
