+++ FrEee.Core.Domain/Processes/Combat/IDamageable.cs
@@ -1,4 +1,4 @@
-﻿using FrEee.Serialization;
+﻿using FrEee.Objects.GameState;
 using FrEee.Utility;
 using FrEee.Utility;
 namespace FrEee.Processes.Combat;
